refactor(radiant-ui-docs): remove playground layer and content codegen
