+++ apps/radiant-ui/src/content/components.ts
@@ -1,5 +1,13 @@
 import { z } from 'zod';
-import type { ComponentCategory } from '@/lib/playground';
+
+export type ComponentCategory =
+	| 'Actions'
+	| 'Forms'
+	| 'Layout'
+	| 'Navigation'
+	| 'Overlays'
+	| 'Feedback'
+	| 'Data display';
 
 /** Public URL prefix for docs pages. */
 export const DOCS_ROOT = '/docs';
