--- apps/radiant-ui/src/lib/playground/index.ts
@@ -1,17 +1,0 @@
-export type {
-	ComponentCategory,
-	ComponentDoc,
-	ComponentGuidanceSection,
-	PlaygroundConfig,
-	PlaygroundControl,
-	PlaygroundScenario,
-	ResolvedPlaygroundState,
-} from './types';
-
-export { booleanControl, numberControl, selectControl, textControl } from './controls';
-export { defineScenario } from './scenario';
-export { definePlayground } from './playground';
-export { defineComponentDoc } from './define-component-doc';
-export { defaultPropsFromControls, resolvePlaygroundState } from './resolve-state';
-export { buildExampleCode, playgroundControlCount } from './build-example-code';
-export { buildPlaygroundExampleCode } from './example-code';
