--- apps/radiant-ui/src/lib/playground/scenario.ts
@@ -1,6 +1,0 @@
-import type { PlaygroundScenario } from './types';
-
-/** Declares a named playground example with optional fixed props and controls. */
-export function defineScenario(scenario: PlaygroundScenario): PlaygroundScenario {
-	return scenario;
-}
